#24 simple sns: 코드 최적화

+++ simple_sns/src/main/java/com/example/sns/controller/PostController.java
@@ -64,7 +64,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId, Authentication authentication) {
+    public Response<Long> likeCount(@PathVariable Integer postId, Authentication authentication) {
         return Response.success(postService.likeCount(postId));
     }
